Docstrings

EcoNetPlot.plot_network — Method

plot_network(A; layout = :umap, tl_axis = :y)

Plot network graph from adjacency matrix. By default, species are positioned according to their trophic levels along the y-axis (can be changed with tl_axis argument). Their position along the x-axis is determined using umap embedding, but other options are possible (see layout).

Argument

A: the adjacency matrix (filled with zeros and ones, or booleans).

Keyword arguments

layout: the type of layout, either 1) :umap which uses umap to compute the embedding, 2) :random which sets random values, or 3) :aligned which sets all values equal (useful for chains).

tl_axis: set the axis of the trophic levels, either :x or :y.

kwargs: keyword argument given to the graphplot function of GraphMakie.

EcoNetPlot.plot_network — Method

plot_network(fw::END.Foodweb; kwargs...)

Plotting directly from a Foodweb object of EcologicalNetworksDynamics.

EcoNetPlot.plot_network — Method

plot_network(m::END.Model; kwargs...)

Plotting directly from a Model object of EcologicalNetworksDynamics.

EcoNetPlot.spread_points — Method

spread_points(vec; min_dist = 0.015)

Space points so they do not overlap.